Don't save models with NaNs.
Leading toward more parameters in the attention blocks and smaller training sequence lengths for now. If things go well, I'll rent an A100 and use a larger training sequence length.

It's maybe too early to be sure, but I"m happy so far as this round of training converges stably. We'll see about how good the results are.