Working toward a final model. It's possible I'll bump into an issue while training is finishing or while testing afterwards, but for right now, I think I've found a balance between stability, speed, and memory constraints.
120 million parameters doesn't sound like a ton, but you have to also remember that this is unlimited (sort of) context.

The big thing I need is to maximize embedding and attention to capture the nuance of tokens and blocks. Unfortunately, it's really tough to train reliably (or quickly) with less than a batch size of 64.

My biggest concern with this version of the model is how well it will handle long messages and whether the 72-token training sequence length will be enough.